Refactor authentication forms: consolidate components and remove unused files. - Moved ForgotPasswordForm, LoginForm, ResetPasswordForm, and SignupForm to their respective files. - Updated import paths in AuthForms component. - Removed old implementations of forms and replaced them with new functional components.
