+++ client/src/App.jsx
@@ -8,7 +8,7 @@ import MyFavorites from "./pages/MyFavorites/MyFavorites";
 import AuthForms from "./components/AuthForms/AuthForms";
 import About from "./pages/About/About";
 import ProtectedRoute from "./components/ProtectedRoute";
-import ResetPasswordForm from "./pages/ResetPassword/ResetPassword";
+import ResetPasswordForm from "./pages/ResetPassword";
 
 function App() {
   return (
