+++ client/src/pages/LoginForm.jsx
@@ -1,0 +1,194 @@
+import { useState, useEffect } from "react";
+import { LogIn, Mail, Eye, EyeOff } from "lucide-react";
+import { UseUser } from "../context/UserContext";
+import AlertMessage from "../components/AlertMessage/AlertMessage";
+import { gif } from "../assets";
+import useFetch from "../hooks/useFetch";
+import { fixUserSkills } from "../util/fixUserSkills";
+
+const LoginForm = ({
+  setLoginSuccessPopup,
+  switchToSignup,
+  switchToForgotPassword,
+}) => {
+  const [loginData, setLoginData] = useState({ email: "", password: "" });
+  const [showPassword, setShowPassword] = useState(false);
+  const [alert, setAlert] = useState({ type: "", message: "" });
+  const { dispatch } = UseUser();
+
+  function handleClearAlert() {
+    setAlert({ type: "", message: "" });
+  }
+
+  function handleLoginResults(data) {
+    const normalizedSkills = fixUserSkills(data.user.skills);
+    const favoriteJobs = Array.isArray(data.user.favorites)
+      ? data.user.favorites.map((job) => ({
+          id: job.id,
+          title: job.title,
+          organization: job.organization,
+          organization_url: job.organization_url,
+          employment_type: job.employment_type,
+          url: job.url,
+          organization_logo: job.organization_logo,
+          display_location: job.display_location,
+          work_mode: job.work_mode,
+          seniority: job.seniority,
+          description_text: job.description_text,
+          date_posted: job.date_posted,
+          travel_time: job.travel_time,
+          least_transfers: job.least_transfers,
+          normalized_description: job.normalized_description,
+        }))
+      : [];
+
+    dispatch({
+      type: "LOGIN",
+      payload: {
+        ...data.user,
+        skills: normalizedSkills,
+        favorites: favoriteJobs,
+      },
+    });
+    setLoginSuccessPopup(true);
+  }
+
+  const { isLoading, error, performFetch } = useFetch(
+    "/users/login",
+    handleLoginResults,
+  );
+
+  useEffect(() => {
+    if (error) {
+      setAlert({ type: "error", message: String(error) });
+      setLoginSuccessPopup(false);
+    }
+  }, [error]);
+
+  const handleChange = (e) => {
+    const { name, value } = e.target;
+    setLoginData({ ...loginData, [name]: value });
+    handleClearAlert();
+  };
+
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+    performFetch({
+      method: "POST",
+      body: JSON.stringify({
+        email: loginData.email,
+        password: loginData.password,
+      }),
+      credentials: "include",
+    });
+  };
+
+  return (
+    <div className="form-card" id="login-form">
+      <form onSubmit={handleSubmit}>
+        <label>
+          Email <span style={{ color: "red", marginLeft: "2px" }}>*</span>
+        </label>
+        <div className="input-wrapper">
+          <input
+            type="email"
+            name="email"
+            placeholder="Enter your email"
+            value={loginData.email}
+            onChange={handleChange} // call handleChange
+            required
+            style={{ paddingRight: "35px" }}
+          />
+          <Mail size={18} className="input-icon-right" />
+        </div>
+        <label>
+          Password <span style={{ color: "red", marginLeft: "2px" }}>*</span>
+        </label>
+        <div className="input-wrapper">
+          <input
+            type={showPassword ? "text" : "password"}
+            name="password"
+            placeholder="Enter your password"
+            value={loginData.password}
+            onChange={handleChange} //call handleChange
+            required
+            style={{ paddingRight: "35px" }}
+          />
+          {showPassword ? (
+            <EyeOff
+              size={18}
+              className="input-icon-right"
+              onClick={() => setShowPassword(false)}
+            />
+          ) : (
+            <Eye
+              size={18}
+              className="input-icon-right"
+              onClick={() => setShowPassword(true)}
+            />
+          )}
+        </div>
+
+        {alert.message && (
+          <AlertMessage type={alert.type} message={alert.message} />
+        )}
+
+        <button type="submit" disabled={isLoading}>
+          {isLoading ? (
+            <>
+              <span>Logging in...</span>
+              <img src={gif.spinner} className="spinner" />
+            </>
+          ) : (
+            <>
+              <LogIn size={16} style={{ marginRight: "5px" }} />
+              Login
+            </>
+          )}
+        </button>
+      </form>
+
+      <p className="switch-text">
+        Don’t have a profile?{" "}
+        <span className="switch-link" onClick={switchToSignup}>
+          Sign Up
+        </span>
+      </p>
+      <p className="forgot-text">
+        <a
+          href="#"
+          className="forgot-link"
+          onClick={(e) => {
+            e.preventDefault(); // prevent page reload
+            switchToForgotPassword();
+          }}
+        >
+          Forgot Password?
+        </a>
+      </p>
+      <p
+        className="gdpr-text"
+        style={{
+          fontSize: "12px",
+          marginTop: "10px",
+          color: "#666",
+          lineHeight: "1.4",
+        }}
+      >
+        By signing up, you agree to the{" "}
+        <a
+          href="https://eur-lex.europa.eu/eli/reg/2016/679/oj/eng"
+          target="_blank"
+          rel="noopener noreferrer"
+          style={{ color: "#0070f3", textDecoration: "underline" }}
+        >
+          GDPR & Privacy Policy
+        </a>
+        . We handle your data according to GDPR rules. Your information is
+        secure and will be deleted if you remove your account.
+      </p>
+    </div>
+  );
+};
+
+export default LoginForm;
